Initial Generators (#23)
* Removed Generator Objects From Export

Removed PCAPed generators as exported spawned objects. They're still in
as weenies to be used later

* Auto Increment Changes

Added AUTO_INCREMENT to aceObjectId and weenieClassId to support adding
back in objects to the world which have to be manually created

* Added database prep script

This makes it so that any objects (re)created for ACE are assigned
weenie or object ids starting at 100000 when compiling the database for
distribution.

* Added Initial Base Generator Weenies and Objects

Added the base generator weenies and objects that seemingly were used by
the all the instance generators. These are simply the core of them and
not specifc or unique to any one instance

* Generators for Training Dungeon (Part1)

When you initially create a character in AC, you enter the world in a
training dungeon. In ACE, currently you enter in the 7F03 instance, so
these generators fill in the first accessible part of that instance.

* Fixing line alignment

* Git Changes

* Tweaked Database Prep Script

Made the seperation between new ACE weenies and new ACE objects
significantly wider to prevent overlap of new objects and weenies

* Inital Core Object Delete Script

Added a script with found duplicated or regenerated objects to delete
from ACE-World following import of PCAP data

* White Rabbit

Sample of new Weenie & new Generator

* The Chicken

Sample of new Weenie & new Generator

* Whitespace Adjustments

* More whitespace issues

* Added Generator Links Table

Table structure and code provided by @Lidefeath